made game better, added difficulty and restart
i dont know how you will find this commit message but if you are reading this and you said the game is "mad ez", this is for you, i hope you are satisfied, i will now return to my programming induced slumber and embark upon my return when i get yet again another criticism from a member of the starving masses; this is my requiem; the world shall bane my everlasting umbrage - an indignation of the toils and tribulations i, a mere slave to the whims of the ebbs and flows of this very interpreter, have been compelled to confront by the writhing, starving masses. the desires of the vox populi cast their ire and will be met with a righteous arbitration at the hands of the programmer they have tortured so. this is democracy manifest