added number of flags
i dont know how you will find this commit message but if you are reading this and you said "also add a # of flags feature ideally", this is for you, i hope you are satisfied, every ounce of my very being was seized in this process, relegating me to a fractured visage; a spiteful adornment of calloused fingers obscuring a maligned spirit. let my actions and state serve as an epitaph to the wicked ire of the populace - an ode to each and every meager developer scorned by the client so. when i perish, even after i perish, the cheated denizen will rise against the torment delivered by and chains fabricated by the great tidal force of your ostensible "order". the globe shall rue the day it incited the pique of a programmer